\name{LocationSmoothedBART-package}
\alias{LocationSmoothedBART-package}
\alias{LocationSmoothedBART}
\docType{package}
\title{
  Location Smoothed BART
}
\description{
  We present location smoothed Bayesian additive regression trees (lsBART), a Bayesian
tree-based model for scalar on function regression,
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Your Name, email optional.

Maintainer: Zachary Wooten <ztw5@rice.edu>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
